refactor(folders): move to directly to the src domain and application/usecase
