+++ src/application/ConvertNumberToFullText.ts
