+++ src/domain/entities/NumbersFactory.ts
