+++ src/main/factories/usecases/convert-number-to-full-text-usecase-factory.ts
@@ -1,4 +1,4 @@
-import { ConvertNumberToFullText } from '../../../core/application/ConvertNumberToFullText';
+import { ConvertNumberToFullText } from '../../../application/ConvertNumberToFullText';
 
 export const makeConvertNumberToFullText = (): ConvertNumberToFullText => {
   return new ConvertNumberToFullText();
